feat: move decapCMS related components
