+++ src/features/decapCMS/FeaturedPost.astro
@@ -4,7 +4,7 @@ import { Picture } from "astro:assets";
 import { formatDate } from "@js/utils";
 
 import { getSiteContext } from "@js/getSiteContext";
-import { getPostSlug } from "@js/blogUtils";
+import { getPostSlug } from "src/features/decapCMS/core/blogUtils";
 import { getRoute } from "@js/routes";
 
 const { locale, content } = await getSiteContext(Astro.url);
