+++ src/pages/fr/blog/[...page].astro
@@ -1,10 +1,11 @@
 ---
 import type { GetStaticPaths, Page } from "astro";
 import { getSiteContext } from "@js/getSiteContext";
+import { getBlogPosts } from "src/features/decapCMS/core/getBlogPosts";
 import BaseLayout from "@layouts/BaseLayout.astro";
 import CTA from "@components/CTA/CTA.astro";
 import Banner from "src/features/demo/Banner.astro";
-import BlogPreview from "@components/BlogPreview/BlogPreview.astro";
+import BlogPreview from "src/features/decapCMS/BlogPreview.astro";
 import Pagination from "@components/Pagination/Pagination.astro";
 
 export const getStaticPaths = (async ({ paginate }) => {
@@ -20,7 +21,6 @@ export const getStaticPaths = (async ({ paginate }) => {
 }) satisfies GetStaticPaths;
 
 import landingImage from "@assets/images/landing.jpg";
-import { getBlogPosts } from "@js/getBlogPosts";
 
 const { content } = await getSiteContext(Astro.url);
 const { page } = Astro.props as { page: Page };
